+++ packages/core/src/components/ui/Alert/Alert.tsx
@@ -15,7 +15,7 @@ export type Props = AlertProps & {
    */
   icon?: ReactNode
   /**
-   * Add the possibility to make the component dismissible
+   * Enables dismissible feature
    */
   dismissible?: boolean
   /**